// LBB.Web/app/src/app/api/types.gen.ts
export type AuthenticatedResponseDto = {
    email?: string;
    fullName?: string;
    token?: string;
    tenant?: Tenant;
    roles?: Array<string>;
    tokenExpiry?: string;
};
// ...
export type LoginRequest = {
    email: string;
    password: string;
    twoFactorCode?: string | null;
    twoFactorRecoveryCode?: string | null;
};
// ...
export type Tenant = {
    id: string;
    description?: string;
    displayName?: string;
    name?: string;
};
// ...
export type PostLoginData = {
    body: LoginRequest;
    path?: never;
    query?: never;
    url: '/login';
};

export type PostLoginResponses = {
    /**
     * OK
     */
    200: AuthenticatedResponseDto;
};

export type PostLoginResponse = PostLoginResponses[keyof PostLoginResponses];

export type PostRefreshData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/refresh';
};